gpsTestMessage: a binary which connects to server and sends a gps message to the server using command line args. Exercises the entire chain of events that currently exist.
bugs: does not correctly parse the ack sent back from the server. The dataMarshaller is seeing the MessageStatus as an unreagistered class even though testCC parses it correctly and even
if I send a testMessage instead of a gpsMessage. Odd. I'm beginning to not like templates at all and may remove them from the dataMarhaller.